requirements.txt
streamlit>=1.31.0
pandas
numpy
plotly
sqlalchemy
psycopg2-binary
toml
xlsxwriter
requests
google-api-python-client
google-auth